[DDW-1068 Remove import from main on renderer]
